+++ eslint.config.js
@@ -1,0 +1,5 @@
+import baseConfig from "./src/rulesets/base.js";
+
+export default {
+    ...baseConfig,
+};
