+++ src/rulesets/jsdoc.js
@@ -1,7 +1,9 @@
+import jsdoc from "eslint-plugin-jsdoc";
+
 /* eslint sort-keys: ["error", "asc"] */
 /* eslint-disable sort-keys */
 
-module.exports = {
+export default {
     plugins: ["jsdoc"],
     overrides: [
         {
