+++ src/styles/prefer-array-shorthand.js
